feat(core)!: 深度重构渲染引擎（Phase 1-3）
- 结构化错误类型: 新增 error.rs, 替换 Box<dyn Error> → CoreError
- 字体引擎迁移: rusttype → ab_glyph (v0.2.32), 解决内存泄漏
- 多字重字体系统: FontSet { regular, bold }, weight 字段生效
- 四种渲染模式: BottomFrame / GradientFrame / Minimal / Overlay
- 右对齐支持: text_width() 计算后真正右对齐
- Alpha 颜色: #RRGGBBAA 格式支持
- Logo 双线性缩放: 消除锯齿
- 四个内置模板重设计: Classic/Compact/Professional/Overlay

+++ litemark-core/src/image_io.rs
@@ -132,7 +132,7 @@ fn decode_heic_from_bytes(data: &[u8]) -> Result<DynamicImage, Box<dyn std::erro
     // Convert to RGBA format for consistency
     let mut rgba_data = Vec::with_capacity((width * height * 4) as usize);
     let rgb_data = interleaved_plane.data;
-    let stride = interleaved_plane.stride as usize;
+    let stride = interleaved_plane.stride;
 
     for y in 0..height {
         let row_start = y as usize * stride;
